@@ -1,26 +1,26 @@
 using System;
 using System.Globalization;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace Website.Api.ReCaptcha
 {
     public class ReCaptchaResponseModel
     {
-        [JsonProperty("success")]
+        [JsonPropertyName("success")]
         public bool Success { get; set; }
 
         public DateTime ChallengeAt { get; set; }
 
-        [JsonProperty("challenge_ts")]
+        [JsonPropertyName("challenge_ts")]
         public string ChallengeTimestamp
         {
             set => ChallengeAt = DateTime.Parse(value, CultureInfo.CreateSpecificCulture("en-US"));
         }
 
-        [JsonProperty("hostname")]
+        [JsonPropertyName("hostname")]
         public string Hostname { get; set; }
 
-        [JsonProperty("error-codes")]
+        [JsonPropertyName("error-codes")]
         public string[] ErrorCodes { get; set; }
     }
 }
